separate partially_indexed_node
